Change the maximum length of commit messages to up to 71 characters
Many people recommend limiting the length of the first line of the commit
message to up to 50 characters, but I find it is sometimes too tightly
limited to write a brief and well-understood summary of the change.

Github displays up to 71 characters in the list of commit messages before
the message is interrupted with a [...] button.

The new limit of the first line of the commit message is therefore up to
71 characters.